Fix little bugs

+++ app/src/main/java/ir/adicom/androidoldpractice/models/Note.kt
@@ -7,7 +7,7 @@ import java.io.Serializable
 
 @Entity(tableName = "notes_table")
 data class Note(
-    @PrimaryKey(autoGenerate = true) val id: Int?,
+    @PrimaryKey(autoGenerate = true) val id: Long?,
     @ColumnInfo(name = "title") val title: String?,
     @ColumnInfo(name = "note") val note: String?,
     @ColumnInfo(name = "date") val date: String?
